Polyclay-cassandra can now store & retrieve set & mapping typing.
The secret is not to mix thrift & CQL3. We are still using thrift for
the attachments table, but the model persistence is now accomplished
through CQL alone. Query string building & quoting is still a little
messy.

The set type needs further refinement before it's complete, but the
basic text set type works.

Also todo: array fields in Polyclay need to become list<text> in Cassandra.
Or something.